pyproject.toml
[tool.poetry]
name = "tgops"
version = "0.1.0"
description = "Self-hostable Telegram community ops toolkit"
authors = []
license = "MIT"
readme = "README.md"
packages = [{include = "tgops"}]

[tool.poetry.dependencies]
python = "^3.11"
hydrogram = "^0.1"
typer = {extras = ["all"], version = "^0.9"}
rich = "^13.0"
pydantic-settings = "^2.0"
pyyaml = "^6.0"
aiofiles = "^23.0"
httpx = "^0.25"
fastapi = "^0.110"
uvicorn = {extras = ["standard"], version = "^0.27"}

[tool.poetry.group.dev.dependencies]
pytest = "^7.4"
pytest-asyncio = "^0.21"
ruff = "^0.1"
black = "^23.0"
mypy = "^1.0"

[tool.poetry.scripts]
tgops = "tgops.main:app"

[tool.pytest.ini_options]
asyncio_mode = "auto"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
